backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
celery==5.5.3
channels==4.3.1
channels_redis==4.3.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.8.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
kombu==5.5.4
msgpack==1.1.1
mysqlclient==2.2.7
packaging==25.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
